+++ Task_1_1_3/src/main/java/ru/nsu/martynov/AbstractExpression.java
@@ -1,0 +1,17 @@
+package ru.nsu.martynov;
+
+public abstract class AbstractExpression implements Expression{
+    @Override
+    public int hashCode() {
+        return toString().hashCode();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) return true;
+        if (obj == null || getClass() != obj.getClass()) return false;
+
+        Expression exp = (Expression) obj;
+        return this.toString().equals(exp.toString());
+    }
+}
