+++ Task_1_1_3/src/main/java/ru/nsu/martynov/Add.java
@@ -9,7 +9,7 @@ import java.util.Map;
  *   compute its derivative;
  *   evaluate its value.
  */
-public class Add extends Expression {
+public class Add extends AbstractExpression {
     private final Expression left;
     private final Expression right;
 
@@ -56,8 +56,8 @@ public class Add extends Expression {
         // Если оба выражения — числа, вычисляем результат
         if (l instanceof Number && r instanceof Number) {
             return new Number(((Number) l).eval("") + ((Number) r).eval(""));
+        } else {
+            return new Add(l, r); // Вернуть упрощённое сложение
         }
-
-        return new Add(l, r); // Вернуть упрощённое сложение
     }
 }
