+++ Task_1_1_3/src/main/java/ru/nsu/martynov/Number.java
@@ -9,7 +9,7 @@ import java.util.Map;
  *   compute its derivative;
  *   and evaluate its value.
  */
-public class Number extends Expression {
+public class Number extends AbstractExpression {
     private final double constNum;
 
     /**
